Refactor binary resource fetching into operation independent of bootstrap
