+++ etc/spack/defaults/windows/resource.yaml
@@ -1,4 +1,4 @@
-bootstrap-resource:
+resource:
   # If set to false Spack will not fetch bootstrap resources
   enable: true
   # Trusted sources to acquire bootstrap resources from
@@ -6,7 +6,7 @@ bootstrap-resource:
     gpg:
       providers:
         - name: Spack Bootstrap Resource Repo
-          endpoint: https://spack.github.io/windows-bootstrap-resources/resources/gpg4win/gpg4win.tar.gz
+          endpoint: https://spack.github.io/windows-bootstrap-resources/resources/gpg/gpg4win.tar.gz
           sha256: 249ab87bd06abea3140054089bad44d9a5d1531413590576da609142db2673ec
     file:
       providers:
